refactor(upload): move upload module to modular monolith structure
Relocate the upload module from src/upload/ into src/modules/upload/
with controllers/, services/ and dtos/ subfolders, per
software-engineering/04_modular_monolith_architecture.md.

No repositories/ or entities/ folder is created: upload is an S3
adapter that owns no tables, so an empty data-access layer would be
structure for its own sake.

Rename file-dto.ts to file.dto.ts to match the .dto.ts convention used
by every other DTO in the codebase.

PrismaModule stays in the module imports. It looks unused because
UploadService only depends on ConfigService, but the controller's
PermissionsGuard injects PrismaService and is resolved from this
module's injector, so removing it breaks every guarded upload route at
boot. Documented inline and left for the auth module refactor, where
the guard should depend on an auth-owned contract instead.

Behavior is unchanged: routes, guards and DTO shapes are untouched.
