+++ src/app.module.ts
@@ -14,7 +14,7 @@ import { PrismaModule } from './prisma/prisma.module';
 import { QueuesModule } from './queues/queues.module';
 import { RedisModule } from './redis/redis.module';
 import { RoleModule } from './modules/role/role.module';
-import { UploadModule } from './upload/upload.module';
+import { UploadModule } from './modules/upload/upload.module';
 import { UserModule } from './user/user.module';
 import { WebhooksModule } from './webhooks/webhooks.module';
