--- src/upload/upload.module.ts
@@ -1,11 +1,0 @@
-import { Module } from '@nestjs/common';
-import { UploadService } from './upload.service';
-import { UploadController } from './upload.controller';
-import { PrismaModule } from '../prisma/prisma.module';
-
-@Module({
-  imports: [PrismaModule],
-  controllers: [UploadController],
-  providers: [UploadService],
-})
-export class UploadModule {}
